Move all functions to multi-src
