+++ src/en/wcofun/src/eu/kanade/tachiyomi/animeextension/en/wcofun/Wcofun.kt
@@ -1,50 +1,8 @@
 package eu.kanade.tachiyomi.animeextension.en.wcofun
 
-import eu.kanade.tachiyomi.animesource.model.AnimesPage
-import eu.kanade.tachiyomi.animesource.model.SEpisode
 import eu.kanade.tachiyomi.multisrc.wcotheme.WcoTheme
-import eu.kanade.tachiyomi.util.asJsoup
-import okhttp3.Response
 
 class Wcofun : WcoTheme() {
-
     override val name = "Wcofun"
-
     override val baseUrl = "https://www.wcoflix.tv"
-
-    override fun episodeListParse(response: Response): List<SEpisode> {
-        val document = response.asJsoup()
-        return document.select(episodeListSelector()).map { episodeFromElement(it) }
-            // If opening an episode link instead of anime link, there is no episode list available.
-            // So we return the same episode with the title from the page.
-            .ifEmpty {
-                listOf(
-                    SEpisode.create().apply {
-                        setUrlWithoutDomain(response.request.url.toString())
-                        val title = document.select(".video-title").text()
-                        val (name, _) = episodeTitleFromElement(title)
-                        this.name = name
-                    },
-                )
-            }
-    }
-
-    override suspend fun getLatestUpdates(page: Int): AnimesPage {
-        return client.newCall(latestUpdatesRequest(page))
-            .execute()
-            .use { response ->
-                if (page == 1) {
-                    latestUpdatesParse(response).copy(hasNextPage = true)
-                } else {
-                    val document = response.asJsoup()
-
-                    val animes = document.select("div.recent-release:contains(Recently Added) + div > ul > li")
-                        .map { element ->
-                            latestUpdatesFromElement(element)
-                        }
-
-                    return AnimesPage(animes, false)
-                }
-            }
-    }
 }
