+++ src/en/wcostream/src/eu/kanade/tachiyomi/animeextension/en/wcostream/WCOStream.kt
@@ -4,6 +4,7 @@ import eu.kanade.tachiyomi.animesource.model.AnimeFilterList
 import eu.kanade.tachiyomi.animesource.model.SAnime
 import eu.kanade.tachiyomi.multisrc.wcotheme.WcoTheme
 import org.jsoup.nodes.Document
+import org.jsoup.nodes.Element
 
 class WCOStream : WcoTheme() {
 
@@ -13,10 +14,14 @@ class WCOStream : WcoTheme() {
 
     override fun popularAnimeSelector(): String = "div#content div.menu ul > li > a"
     override fun latestUpdatesSelector(): String = "div#content > div > div:has(div.recent-release:contains(Recent Releases)) > div > ul > li"
+    override fun searchAnimeSelector(): String = "div#blog > div.iccerceve"
 
     override fun episodeListSelector() = "div#catlist-listview > ul > li, table:has(> tbody > tr > td > h3:contains(Episode List)) div.menustyle > ul > li"
 
-    override fun searchAnimeSelector(): String = "div#blog > div.iccerceve"
+    override fun latestUpdatesFromElement(element: Element): SAnime {
+        return super.latestUpdatesFromElement(element)
+            .apply { title = title.substringBefore(" Episode").trim() }
+    }
 
     override fun animeDetailsParse(document: Document) = SAnime.create().apply {
         document.selectFirst("div.video-title a")?.text()?.let { title = it }
